Use single config array for WorldPay configurations
This commit refactors how the driver is configured. Each merchant code can have different values so these need reflected in the config array.

Merchant codes are differentiated by the currency they support as well as whether they are used for customer present/not present transactions.

The driver is updated to choose the most appropriate configuration for a given currency/customer present scenario and expose these via public methods for consumption by the driver internally as well as by any external party, if needed.